Added compact schemes to improve frequency analysis
A 6th order compact scheme was added to compute the time-derivative for the instantaneous frequency (it was shwon to be of minor importance, so that the 1st order forward scheme from `np.diff` as kept as default). Also, a 6th order compact filter was included to eliminate the ringing that appears in the imaginary part of the Hilbert transformed signal. The usage of such high-frequency, low-strength filter was shown to improve the time-frequency analysis of the signal.